Add files via upload
The Pantheon is a former Roman temple, now a church, in Rome, Italy, on the site of an earlier temple commissioned by Marcus Agrippa during the reign of Augustus (27 BC – 14 AD). It was completed by the emperor Hadrian and probably dedicated about 126 AD. Its date of construction is uncertain because Hadrian chose not to inscribe the new temple but rather to retain the inscription of Agrippa's older temple, which had burned down. We have tried to represent this amazing in 3D using OpenGl.